requirements.txt
Flask==2.0.0
Werkzeug==2.0.0
stripe==2.56.0
python-dotenv==0.19.0
